# pkg/bind/openapi/api.yaml
openapi: "3.0.0"

info:
  title: "Bind Wallet API"
  description: "API to create CVU Transfers and QR payments"
  version: "1.0.0"

servers:
  - url: "{server}/walletentidad-operaciones/v1/api/v1.201"
    variables:
      server:
        default: https://api.bindpagos.com.ar
        enum:
          - https://gw-staging-qrbind.epays.services
          - https://api.bindpagos.com.ar

security:
  - bearerAuth: []

paths:
  /transferir:
    post:
      summary: "Creates an outbound transfer"
      operationId: "createTransfer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createTransferRequest"
      responses:
        "200":
          description: "Transfer successfully created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/createTransferResponse"
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/apiError"
        "401":
          description: "Unauthorized"


components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes
  schemas:
    createTransferRequest:
      type: object
      properties:
        cvuOrigen:
          type: string
          description: "CVU de la cuenta origen desde la cual se debitará los fondos para realizar la transferencia."
        cbu_cvu_destino:
          type: string
          description: "CBU/CVU de la cuenta destino a la cual se le deben acreditar los fondos transferidos. Puede omitirse si se envía alias."
        cuit_destino:
          type: string
          description: "CUIL/CUIT del titular de la cuenta destino a la cual se le deben acreditar los fondos transferidos. Este campo es opcional pero si se envía pueden aplicarse reglas de monitoreo transaccional sobre el destinatario de las transferencias."
        alias_destino:
          type: string
          description: "Alias de la cuenta destino a la cual se le deben acreditar los fondos transferidos. Puede omitirse si se envía cbu_cvu_destino."
        importe:
          type: number
          description: "Valor del importe a transferir."
        referencia:
          type: string
          description: "Descripción de referencia de la transferencia."
        concepto:
          type: string
          description: "Concepto de la transferencia."
        emails:
          type: array
          description: "Array de strings con lista de correos electrónicos de los destinatarios a los cuales se les enviará un comprobante de la transferencia. El comprobante de la transferencia es enviado por nuestro banco sponsor y no podemos modificar el diseño con el que es enviado."
        idExterno:
          type: string
          maxLength: 50
          description: "Identificador o código externo de la entidad informado opcionalmente para ser relacionado como atributo adicional de la operación. Puede utilizarse para luego realizarse consultas y otras acciones referenciandose a este id externo."
      required:
        - cvuOrigen
        - importe
    createTransferResponse:
      type: object
      properties:
        operacionId:
          type: integer
        operacionIdExterno:
          type: string
        estadoExterno:
          type: string
          description: "Descripción del estado de la transferencia informado por nuestro procesador externo (en este caso, el banco). El estado UNKNOWN se produce cuando hubo un error de comunicación con algún sistema externo, el mismo tratará de actualizarse durante el día. En el caso de una transferencia a un CVU, inicialmente quedan en estado IN_PROGRESS, este estado se intentará resolver durante el día o al final del mismo cuando se concilien todas las operaciones. El estado UNKNOWN_FOREVER indica que la operación no pudo resolver el estado, y no se va a reintentar actualizar en forma automática. Los estados COMPLETED o FAILED son estado finales que indican que la operación se completó o falló."
          enum:
            - UNKNOWN
            - IN_PROGRESS
            - UNKNOWN_FOREVER
            - COMPLETED
            - FAILED
        estadoId:
          type: integer
          description: "Identificador del estado de la operación. Este es el estado que importa en nuestro sistema."
          enum:
            - 1 # A procesar
            - 2 # Aprobada
            - 3 # Rechazada
            - 4 # A consultar
            - 5 # Auditar
            - 6 # Devuelta
            - 7 # Devuelta parcialmente
        origenCuentaId:
          type: integer
          decription: "Identificador de la cuenta de la que se debitaron los fondos."
        coelsaId:
          type: string
          description: "Identificador de Coelsa por esta transferencia. Es el identificador más importante para intercambiar consultas y reclamos con otros bancos, billeteras o proveedores de cuenta."
        fechaInicio:
          type: string
          format: "date-time"
          description: "Fecha y hora en que finalizó la transferencia en el banco."
        fechaFin:
          type: string
          format: "date-time"
          description: "Fecha y hora en que finalizó la transferencia en el banco."
        fechaNegocio:
          type: string
          format: "date-time"
          description: "Fecha y hora en que se instruyó la creación de la transferencia."
        importe:
          type: number
          description: "Valor del importe de la transferencia."
        cvuOrigen:
          type: string
          description: "CVU desde el que se originó la operación."
        referencia:
          type: string
          description: "Referencia de la operación."
        concepto:
          type: string
          description: "Concepto de la operación."
        cvuCbuContraparte:
          type: string
          description: "CBU/CVU de la cuenta destino a la cual se le deben acreditar los fondos transferidos."
        aliasContraparte:
          type: string
          description: "Alias de la cuenta destino a la cual se le deben acreditar los fondos transferidos."
        nombreContraparte:
          type: string
          description: "Nombre del titular de la cuenta destino a la cual se le deben acreditar los fondos transferidos."
        cuitCuilContraparte:
          type: string
          description: "CUIT/CUIL del titular de la cuenta destino a la cual se le deben acreditar los fondos transferidos"
        comprobanteId:
          type: integer
          description: "Identificador del comprobante creado asociado a esta operación. Es el comprobante de débito del saldo."
        comprobanteDevolucionId:
          type: integer
          description: "Identificador del comprobante de la devolución, si corresponde. Este comprobante se crearía eventualmente si la transferencia fue creada pero finalmente su estado es fallido en el banco. Entonces, la transferencia saliente tendrá un comprobante de débito original para retener el saldo y luego tendrá un nuevo comprobante de crédito para devolver el saldo a la cuenta por la transferencia que finalmente no se realizó."
        mensajeAdicional:
          type: string
          description: "Descripción adicional del estado enviado por el procesador (en este caso, el banco)."
        esTransferenciaInterna:
          type: boolean
          description: "Indica si la transferencia es interna. Es decir, si la transferencia se realizó entre cuentas de la misma entidad."
        estaFinalizada:
          type: boolean
          description: "Indica si la operación adquirió un estado definitivo que ya no cambiará."
        estaRechazada:
          type: boolean
          decription: "Indica si la transferencia fue rechazada definitivamente."
        estaAAuditar:
          type: boolean
          description: "Indica si la transferencia está en un estado no definitivo que requiere alguna acción para ser resuelto o no. Que esté pendiente de ser auditada significa que será incluída en nuestro proceso de conciliación automática o será analizada por nuestro equipo de administración para determinar un estado definitivo."
        estaPendiente:
          type: boolean
          description: "Indica si la transferencia está pendiente o no. Que esté pendiente significa que aún no se ha intentado instruir en nuestro procesador externo."
    apiError:
      type: object
      properties:
        eventId:
          type: string
        detalle:
          type: string
        correlationId:
          type: string
        errores:
          type: array
          items:
            type: object
            properties:
              codigo:
                type: string
              titulo:
                type: string
              detalle:
                type: string